% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tivis.r-package.R
\docType{package}
\name{tivis.r-package}
\alias{tivis.r}
\alias{tivis.r-package}
\title{tivis.r: Read Diaspective Vision TIVITA Hyperspectral Recordings}
\description{
Reads hyperspectral imaging data written by the Diaspective Vision TIVITA Suite, used for intraoperative tissue and perfusion imaging. Parses the binary 'SpecCube.dat' container (big-endian float32 samples in band-interleaved-by-pixel order behind a 12-byte dimension header), the German-language acquisition metadata log, and the parameter images the Suite exports alongside each recording (oxygenation, NIR perfusion, tissue haemoglobin and water indices). Requires no vendor SDK and contains no compiled code. Companion to 'cuvis.r', which covers Cubert snapshot cameras, and a data source for 'hyperspectR'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cttir/tivis.r}
  \item \url{https://cttir.github.io/tivis.r/}
  \item Report bugs at \url{https://github.com/cttir/tivis.r/issues}
}

}
\author{
\strong{Maintainer}: R. Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

Authors:
\itemize{
  \item R. Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})
}

}
\keyword{internal}
