% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tivis-measurement.R
\name{tivis_parse_timestamp}
\alias{tivis_parse_timestamp}
\title{Parse a TIVITA Filename Timestamp}
\usage{
tivis_parse_timestamp(x)
}
\arguments{
\item{x}{Character vector of file or directory names.}
}
\value{
A \code{POSIXct} vector, \code{NA} where no timestamp is present.
}
\description{
TIVITA names every artefact of a recording \verb{YYYY_MM_DD_HH_MM_SS_<what>}.
This extracts the acquisition time from such a name.
}
\examples{
tivis_parse_timestamp("2019_11_25_13_29_24_SpecCube.dat")

}
